fetchbot-platform/bots/network-bot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-nmap==0.7.1
dnspython==2.4.2
httpx==0.25.2
pydantic==2.5.2
aiofiles==23.2.1
